Revert change counters to 64 bit

+++ test/perf/core/queue_test.cpp
@@ -19,7 +19,7 @@ namespace core {
 
     void BM_spsc_queue_int(benchmark::State& state) {
         static spsc_queue<std::int32_t, size> queue;
-        std::int64_t count = 0;
+        std::int32_t count = 0;
 
         for (auto _ : state) {
             if (state.thread_index == 0) {
@@ -77,7 +77,7 @@ namespace core {
 	
     static void BM_spsc_queue_s512(benchmark::State& state) {
         static spsc_queue<s512, size> queue;
-        std::int64_t count = 0;
+        std::int32_t count = 0;
 		s512 item = {{0,1,2,3,4,5,6,7}};
 
         for (auto _ : state) {
@@ -105,8 +105,8 @@ namespace core {
     static void BM_spinlock_queue_s512(benchmark::State& state) {
 		s512 item = {{0,1,2,3,4,5,6,7}};
 
-        std::int64_t push_count = 0;
-        std::int64_t pop_count = 0;
+        std::int32_t push_count = 0;
+        std::int32_t pop_count = 0;
 
         for (auto _ : state) {
             if (state.thread_index == 0) {
@@ -138,7 +138,7 @@ namespace core {
         std::queue<std::int32_t> queue;
         spinlock spin;
 
-        std::int64_t count = 0;
+        std::int32_t count = 0;
 
         for (auto _ : state) {
             while (count < state.range(0)) {
@@ -155,7 +155,7 @@ namespace core {
     static void BM_spsc_queue_push(benchmark::State& state) {
         spsc_queue<std::int32_t, 1 << 16> queue;
 
-        std::int64_t count = 0;
+        std::int32_t count = 0;
 
         for (auto _ : state) {
             while (count < state.range(0)) {
@@ -174,7 +174,7 @@ namespace core {
         std::queue<std::int32_t> queue;
         spinlock spin;
 
-        std::int64_t count = 0;
+        std::int32_t count = 0;
 		
 		while (count < state.range(0)) {
             queue.push(count);
@@ -196,7 +196,7 @@ namespace core {
     static void BM_spsc_queue_pop(benchmark::State& state) {
         spsc_queue<std::int32_t, 1 << 16> queue;
 
-        std::int64_t count = 0;
+        std::int32_t count = 0;
 		
 		while (count < state.range(0)) {
             auto res = queue.try_push(count);
